Fix compilation errors in HomeViewModel, HomeView, and RoundActiveView
- Remove unnecessary 'await' keywords from non-async function calls in HomeViewModel.swift
- Remove unnecessary 'await' keywords from non-async function calls in HomeView.swift
- Fix implicit coercion warning by removing redundant Double cast in RoundActiveView.swift

All compilation errors should now be resolved.

+++ RoundsVersion2/Views/RoundActiveView.swift
@@ -1249,7 +1249,7 @@ struct RoundActiveView: View {
                 "statistics.doubleBogeys": FieldValue.increment(Int64(stats["doubleBogeys"] as? Int ?? 0)),
                 "statistics.lastPlayed": FieldValue.serverTimestamp(),
                 "statistics.bestScore": totalScore, // This should be compared with existing best
-                "statistics.averageScore": stats["averageScore"] as? Double ?? 0.0 as Double
+                "statistics.averageScore": stats["averageScore"] as? Double ?? 0.0
             ])
             
             // Add to player's match history
